+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- semantic tokens support for syntax highlighting
+
 ## [2.5.4] - 2026-04-09
 
 ### Changed